refactor(agents, tasks): refactor tasks into a package
