+++ agents_infra/agents_infra/tasks/task_factory.py
@@ -1,0 +1,33 @@
+from .gathering_tasks import (CollectAndSendMetricsTask,
+                              CollectAndSendStatusTask)
+
+
+def task_factory(task_config, agent, supervisor=None):
+    """Task factory to create task instances based on the configuration.
+
+    Args:
+        task_config (dict): A dictionary containing task configuration.
+        agent (ServerAgent): An instance of the agent that
+            the task will operate on.
+        supervisor (AgentSupervisor, optional): A supervisor instance,
+            if a task needs it. Defaults to None.
+
+    Raises:
+        ValueError: If the task name is unknown.
+
+    Returns:
+        BaseTask: A task instance based on the configuration.
+    """
+    name = task_config['name']
+    endpoint = task_config['endpoint']
+    interval = task_config['interval']
+    if name == 'collect_and_send_status':
+        return CollectAndSendStatusTask(agent, endpoint, interval, supervisor)
+    elif name == 'collect_and_send_metrics':
+        return CollectAndSendMetricsTask(agent, endpoint, interval, supervisor)
+    # elif name == "fetch_command":
+    #     return FetchCommandTask(agent, endpoint, interval, supervisor)
+    # elif name == "execute_command":
+    #     return ExecuteCommandTask(agent, endpoint, interval, supervisor)
+    else:
+        raise ValueError('Unknown task name: %s' % name)
